+++ Source/nthread.h
@@ -7,7 +7,7 @@ extern DWORD gdwMsgLenTbl[MAX_PLRS];
 extern DWORD gdwDeltaBytesSec;
 extern BOOLEAN nthread_should_run;
 extern DWORD gdwTurnsInTransit;
-extern int glpMsgTbl[MAX_PLRS];
+extern uintptr_t glpMsgTbl[MAX_PLRS];
 extern unsigned int glpNThreadId;
 extern int turn_upper_bit;
 extern BOOLEAN sgbThreadIsRunning;
